Second pass on the bft actor.
