+++ node/actors/bft/src/replica/commit.rs
@@ -1,43 +1,59 @@
 //! Handler of a ReplicaCommit message.
-
 use super::StateMachine;
 use crate::metrics;
 use std::collections::HashSet;
-use zksync_concurrency::{ctx, metrics::LatencyHistogramExt as _};
-use zksync_consensus_network::io::{ConsensusInputMessage, Target};
+use zksync_concurrency::{ctx, error::Wrap, metrics::LatencyHistogramExt as _};
 use zksync_consensus_roles::validator;
 
-/// Errors that can occur when processing a "replica commit" message.
+/// Errors that can occur when processing a ReplicaCommit message.
 #[derive(Debug, thiserror::Error)]
 pub(crate) enum Error {
     /// Message signer isn't part of the validator set.
-    #[error("Message signer isn't part of the validator set (signer: {signer:?})")]
+    #[error("message signer isn't part of the validator set (signer: {signer:?})")]
     NonValidatorSigner {
         /// Signer of the message.
         signer: Box<validator::PublicKey>,
     },
     /// Past view or phase.
-    #[error("past view/phase (current view: {current_view:?}, current phase: {current_phase:?})")]
+    #[error("past view (current view: {current_view:?})")]
     Old {
         /// Current view.
         current_view: validator::ViewNumber,
-        /// Current phase.
-        current_phase: validator::Phase,
     },
-    /// The processing node is not a lead for this message's view.
-    #[error("we are not a leader for this message's view")]
-    NotLeaderInView,
-    /// Invalid message.
-    #[error("invalid message: {0:#}")]
-    InvalidMessage(#[source] validator::ReplicaCommitVerifyError),
+    /// Duplicate signer.
+    #[error("duplicate signer (current view: {current_view:?}, signer: {signer:?})")]
+    DuplicateSigner {
+        /// Current view.
+        current_view: validator::ViewNumber,
+        /// Signer of the message.
+        signer: Box<validator::PublicKey>,
+    },
     /// Invalid message signature.
     #[error("invalid signature: {0:#}")]
     InvalidSignature(#[source] anyhow::Error),
+    /// Invalid message.
+    #[error("invalid message: {0:#}")]
+    InvalidMessage(#[source] validator::ReplicaCommitVerifyError),
+    /// Internal error. Unlike other error types, this one isn't supposed to be easily recoverable.
+    #[error(transparent)]
+    Internal(#[from] ctx::Error),
+}
+
+impl Wrap for Error {
+    fn with_wrap<C: std::fmt::Display + Send + Sync + 'static, F: FnOnce() -> C>(
+        self,
+        f: F,
+    ) -> Self {
+        match self {
+            Error::Internal(err) => Error::Internal(err.with_wrap(f)),
+            err => err,
+        }
+    }
 }
 
 impl StateMachine {
-    /// Processes `ReplicaCommit` message.
-    pub(crate) fn process_replica_commit(
+    /// Processes a ReplicaCommit message.
+    pub(crate) async fn on_commit(
         &mut self,
         ctx: &ctx::Ctx,
         signed_message: validator::Signed<validator::ReplicaCommit>,
@@ -55,25 +71,21 @@ impl StateMachine {
             });
         }
 
-        // If the message is from the "past", we discard it.
-        // That is, it's from a previous view or phase, or if we already received a message
-        // from the same validator and for the same view.
-        if (message.view.number, validator::Phase::Commit) < (self.view, self.phase)
-            || self
-                .replica_commit_views
-                .get(author)
-                .is_some_and(|view_number| *view_number >= message.view.number)
-        {
+        // If the message is from a past view, ignore it.
+        if message.view.number < self.view {
             return Err(Error::Old {
                 current_view: self.view,
-                current_phase: self.phase,
             });
         }
 
-        // If the message is for a view when we are not a leader, we discard it.
-        if self.config.genesis().view_leader(message.view.number) != self.config.secret_key.public()
-        {
-            return Err(Error::NotLeaderInView);
+        // If we already have a message from the same validator for the same view, ignore it.
+        if let Some(&view) = self.commit_views_cache.get(author) {
+            if view == message.view.number {
+                return Err(Error::DuplicateSigner {
+                    current_view: self.view,
+                    signer: author.clone().into(),
+                });
+            }
         }
 
         // ----------- Checking the signed part of the message --------------
@@ -89,7 +101,7 @@ impl StateMachine {
 
         // We add the message to the incrementally-constructed QC.
         let commit_qc = self
-            .commit_qcs
+            .commit_qcs_cache
             .entry(message.view.number)
             .or_default()
             .entry(message.clone())
@@ -98,21 +110,21 @@ impl StateMachine {
         // Should always succeed as all checks have been already performed
         commit_qc
             .add(&signed_message, self.config.genesis())
-            .expect("Could not add message to CommitQC");
+            .expect("could not add message to CommitQC");
 
         // Calculate the CommitQC signers weight.
         let weight = self.config.genesis().validators.weight(&commit_qc.signers);
 
         // Update commit message current view number for author
-        self.replica_commit_views
+        self.commit_views_cache
             .insert(author.clone(), message.view.number);
 
         // Clean up commit_qcs for the case that no replica is at the view
         // of a given CommitQC
         // This prevents commit_qcs map from growing indefinitely in case some
         // malicious replica starts spamming messages for future views
-        let active_views: HashSet<_> = self.replica_commit_views.values().collect();
-        self.commit_qcs
+        let active_views: HashSet<_> = self.commit_views_cache.values().collect();
+        self.commit_qcs_cache
             .retain(|view_number, _| active_views.contains(view_number));
 
         // Now we check if we have enough weight to continue.
@@ -120,36 +132,29 @@ impl StateMachine {
             return Ok(());
         };
 
-        // ----------- Update the state machine --------------
-        let now = ctx.now();
-        metrics::METRICS
-            .leader_commit_phase_latency
-            .observe_latency(now - self.phase_start);
-        self.view = message.view.number.next();
-        self.phase = validator::Phase::Prepare;
-        self.phase_start = now;
+        // ----------- We have a QC. Now we process it. --------------
 
-        // ----------- Prepare our message and send it. --------------
-
-        // Consume the incrementally-constructed QC for this view.
-        let justification = self
-            .commit_qcs
+        // Consume the created commit QC for this view.
+        let commit_qc = self
+            .commit_qcs_cache
             .remove(&message.view.number)
             .unwrap()
             .remove(message)
             .unwrap();
 
-        // Broadcast the leader commit message to all replicas (ourselves included).
-        let output_message = ConsensusInputMessage {
-            message: self
-                .config
-                .secret_key
-                .sign_msg(validator::ConsensusMsg::LeaderCommit(
-                    validator::LeaderCommit { justification },
-                )),
-            recipient: Target::Broadcast,
-        };
-        self.outbound_pipe.send(output_message.into());
+        self.process_commit_qc(ctx, &commit_qc)
+            .await
+            .wrap("process_commit_qc()")?;
+
+        // Metrics.
+        let now = ctx.now();
+        metrics::METRICS
+            .leader_commit_phase_latency
+            .observe_latency(now - self.phase_start);
+        self.phase_start = now;
+
+        // Start a new view.
+        self.start_new_view(ctx, message.view.number.next());
 
         Ok(())
     }
